Add solver params; update main function API
Added fields to the solver structs where needed (or, at least, indicated with a TODO that fields should be added instead of leaving the struct trivial). Updated the main function API to assert out of the gate that the matrix is square and to copy to avoid shared mutability. Also made the DEFAULT_SOLVER behavior more elegant with a default arg. Finally, corrected the (unfinished) approach docstring in types.jl.