Finish API design; complete Heuristic functionality (#8)
This PR finishes the API design (splitting up the different solver categories into different submodules) and uses this to finish functionality for the Heuristic submodule, with Cuthill-McKee and Reverse Cuthill-McKee now implemented. Docs and tests remain to be completed, however.

A minor change was also made to the README, simply adding a row for static analysis information to the project summary table. Additionally, JET static analysis was modified to allow for some reports now that development has started, so long as the total number stays under 30. (This is standard practice.) Finally, docs/src/index.md was modified to include documentation on the MatrixBandwidth submodules (one for each solver category).