Finish documentation for Heuristic module
Finish all docstrings (including doctests and examples) for the Heuristic module, along with several References. Modify the overall formatting and design within src/ to be consistent with this. Also change the Documenter website accordingly.

Note that documentation of some core MatrixBandwidth names themselves (e.g., minimize_bandwidth) is yet to be completed.