Fix verb ellipsis; update MBMP defn; undo bold emphasis (#22)
First and foremost, this PR fixes an erroneous ellipsis of the verb in the README, the Documenter homepage, and the MatrixBandwidth module docstring. I was very silly and forgot that ellipsis of the verb is only grammatically correct when both conjucts are of the same syntactic category and structure; hence, I added an 'are' after 'thus' in 'exact methods ... are exponential in time complexity and thus [are] only feasible for relatively small matrices'.

Second, this PR also updates the wording of the definition of the matrix bandwidth minimization problem (MBMP) in the aforementioned documentation. Nothing grammatically wrong or even vague--I just wanted to make it flow more nicely.

Finally, I've ultimately decided against making 'at most [k]' and 'at least [k]' bold instead of italic (for emphasis) as was changed in #20, so this PR effectively reverses that one as well.