Finish MB-ID logic
This commit marks the completion of the MB-ID solver logic (also minor polishes remain for clarity/efficiency). In figuring out the MB-ID logic, we also saw that we needed to set the diagonal entries of A_bool in the minimize_bandwidth core function to false; this change was implemented. We also imported the _symmetrize helper (which was extracted from the reverse Cuthill-McKee logic and placed into src/utils.jl for common use) in each submodule (Exact, Heuristic, and Metaheuristic).