MB-ID solver; other minor fixes (#26)
This PR finishes the logic for the MB-ID solver in the MatrixBandwidth.Exact submodule. We also note that the symmetrization of the input matrix originally contained within the reverse Cuthill&ndash;McKee solver is again needed, so we extract the logic into a _symmetrize helper function in src/utils.jl for common use, importing it in all three submodules (MatrixBandwidth.Exact, MatrixBandwidth.Heuristic, and MatrixBandwidth.Metaheuristic). (Looking at issue #25, however, we may change this part of the API anyway, requiring symmetric input in the first place and letting users pick their own symmetrization strategy, at least for the nonzero support.)

We also change the expected date for our first stable release (in the documentation and the README) from late June 2025 to early July, given several minor delays in development.

(Note that unit tests and documentation for the MB-ID solver are not yet done; we shall address this in a subsequent PR. We also start briefly on the logic for the MB-PS solver but leave it largely unfinished.)