Add functionality for recognition as in #5 (#28)
This PR restructures the whole package to add functionality for matrix bandwidth recognition in addition to minimization, as discussed in issue #4.

NOTE: Some sections are still 'skeleton templates' with nothing of substance--the focus of this PR is to handle the API-breaking restructuring, not to add all the logic.

WARNING: This is a highly API-breaking PR. Not only can backwards compatibility not be assumed--it is, indeed, to be assumed that any aspects of the new API are NOT backwards compatible.