Add docstrings; fix test info
This PR adds docstrings to the completed Del Corso-Manzini algorithms (both algorithms, in both Minimization and Recognition) as well as to the Caprara-Salazar-Gonzalez algorithm. Complexity analysis is performed for DCM but not yet for CSG. Examples are forthcoming for both.

It also identifies some room for optimization in the DCM with PS algorithm and notes an error in the original paper (to explain the mismatch between implementation and source).

Finally, it also fixes #41, making sure that proper info is printed for each test group while running unit tests.

(A slight change is also made to the Cuthill-McKee and reverse Cutill-McKee documentation to stay consistent with the new DCM/CSG docstrings, but it is very minor.)