Add brute force algos; incorp. into tests (#45)
We add both a solver and a decider for the brute-force approach to bandwidth minimization/recognition and use this to write effective tests for the DCM algorithms. We also tweak the CM/RCM tests slightly, reducing the number of iterations so they run in a more reasonable time frame.

Additionally, tests are created for the top-level core API of MatrixBandwidth as well.

Finally, a few typos/poor wordings are fixed throughout the document.

(Resolves #42.)