Replace some TODO's with errors (#51)
In critical parts of the code (e.g., mandatory method definitions for specific solvers/deciders), we replace 'TODO: Implement's with actual thrown errors. This is particularly important in stuff like CSG or GPS, which are default algorithms for recognition and minimization, so users are clear on the fact that this just has not been done yet. This pattern is clearly indicated in the 'Algorithms' section of the README (and documentation website) as well.

(As another minor related change, we also add square brackets around some other TODO's in docstrings.)