Finalize DCM comments; improve DCM compat check
We finalize the inline comments for DelCorsoManzini and DelCorsoManziniWithPS (both in Minimization.Exact and in Recognition). Additionally, we find during the process room for optimization (AND greater functional-style clarity) in the _dcm_is_compatible helper function, making the appropriate edits.

Finally, it is noticed that several indexing + reallocs throughout the source code could more efficiently be replaced with views; this is changed.