Move conn. comp. to heuristic utils; improve lazy mapping
This PR, in preparation for the implementation of Gibbs-Poole-Stockmeyer, moves the connected component logic (originally in the file for Cuthill-McKee) into the utils file for the Minimization.Heuristic module. It also does some preliminary setup for the GPS algorithm, improves the clarity/efficiency of the CM algorithm (making some imperative parts more functional), and finds a few other bits of source code in the package that benefit from lazier mapping with Iterators.map as opposed to simply map.