Add utils tests/clarify core examples
This PR nearly finishes the utils test suite, adding tests for all utility functions at the root level of the package. (The only things missing now are a few edge cases for random_banded_matrix.)

Additionally, the TODO's in the Examples sections of the minimize_bandwidth and has_bandwidth_k_ordering are clarified, pointing users to the Examples sections of specific solvers/deciders instead while we wait for all solvers and deciders to be implemented.