Refine readme/docs examples (#61)
In the README/Documenter docs examples, we seed the example to ensure reproducibility. We also show how minimize_bandwidth and has_bandwidth_k_ordering have default algorithms in case no algorithm is specified, noting that the default algorithm for the latter (GPS) is not yet implemented. Finally, we indicate in the listed algorithms which ones are still under development.