Prepare for first stable release, v0.1.0 (#64)
We have decided to skip straight to releasing v0.1.0, as the bulk of the library is already functional and tested with remaining TODO's clearly indicated and documented.

Here we edit descriptions/status in README.md and docs/src/index.md accordingly, also updating the listed version from v0.1.0-beta to v0.1.0 in both of the aforementioned files plus Project.toml.