Edit docs+CITATION.bib; optimize bandwidth; clarify empty reducing
This PR fixes #68 by (1) replacing all instances of 'index' in source code with 'idx' for consistency and (2) avoiding reallocating A_bool with _offdiag_nonzero_support in the bandwidth function. It also switches the generator comprehension in bandwidth for an Iterators.map call and clarifies via inline comments why, in the case that indices is empty, we avoid reducing over an empty collection in maximum. (This same clarification is added to the source code for the Del Corso–Manzini decider in the Recognition submodule when the same strategy is used.)

Moreover, we make a few minor changes to the docs--README.md, docs/src/index.md (for the Documenter site), and the docstring for the MatrixBandwidth module. References to the Documenter site URL are made more general (instead of adding 'dev/' or 'stable/' to the end of said URLs, we just leave it as 'https://luis-varona.github.io/MatrixBandwidth.jl/', which is more idiomatic and should automatically redirect to the stable documentation anyway. In the README and index, we also change our target date for completion of core API development from 'late August 2025' to September 2025. (This does not, of course, include our goals of adding support for sparse matrices and graph formats.)

Finally, we change the BibTeX key in CITATION.bib from 'Var2025' to 'Var25'.