Add profile function
This PR adds the profile function to compute the original profile of a matrix prior to any reordering. It updates the CM/RCM/GPS docstrings accordingly (pointing to this new function's documentation whenever matrix profile is referenced) and elaborates in the MatrixBandwidth module docstring on this and bandwidth as additional core exports of the package. The README and Documenter index are also updated accordingly.

Some other minor changes/fixes include reformatting entries in refs.bib, fixing a rendering error in the dcm_ps_optimal_depth docstring, and fixing some code blocks enclosed in double backticks that should have been enclosed in single backticks.

(Resolves #62.)