Finish SGS solver outline; fix CM typo
This PR finishes the solver logic for Saxe-Gurari-Sudborough, which calls the Recognition decider (not yet done) as a subroutine. (Hence, the solver doesn't actually work yet, but it does not need to be modified any further, it just needs the decider logic to be completed.)

Additionally, we identify a typo in the documentation for Cuthill-McKee.