Update node finders and unit tests
This PR adds 'bi_criteria_node_finder' (an improvement upon 'pseudo_peripheral_node_finder' as a new node finder for the heuristic solvers, changing the 'DEFAULT_NODE_FINDER' constant from the latter to the former in the process.

It also improves unit tests for the heuristic solvers with more edge cases and scenarios (including the use of custom node finders). At the same time, it reduces the number of iterations and cases in some of the unit tests (not just for the heuristic solvers) to cut down on runtime without affecting coverage.

(Resolves #44.)