Finish SGS; rename '_requires_symmetry' (#123)
This PR renames the '_requires_symmetry' internal function (used for input validation) to '_requires_structural_symmetry'. It also completes the implementation of the Saxe–Gurari–Sudborough algorithm (both the minimization solver and the recognition decider), although some documentation is still pending. (Unit tests are all complete for SGS, though.)

On a secondary note, it also fixes some test names in the Del Corso–Manzini recognition algorithm test suite ('Bandwidth < k' was meant to be 'Bandwidth > k', and 'Bandwidth ≥ k' was meant to be 'Bandwidth ≤ k').