Increase number of unit tests (#130)
Previous PRs #112, #113, #118, and #119 reduced the number of iterations and cases in some of the unit tests to cut down on runtime without affecting coverage. This was largely due to a slowdown from thousands of deprecation warnings after the bump to DataStructures v0.19 compat while maintaining backwards compatibility with v0.18. Since #129 fixed this, we now revert back to the old numbers of test iterations and matrix orders, possibly being more robust.

We also reduce the 'timeout-minutes' attribute in our CI.yml workflow back to 60 from 180.