Refactor import structure; update exports
This PR refactors the import structure in submodules to do, e.g., 'using MatrixBandwidth' in favor of a bunch of 'import ..A, ..B, ..C' statements, similarly to how Graphs.jl does it. (On a related note, it also reformats the export blocks in each submodule to be more in cohesive, again following the style of Graphs.jl.)

Additionally, it removes the leading underscore from many (but not all) top-level utility functions (also renaming '_offdiag_nonzero_support' to 'offdiag_nz_support'), adding docstrings to all of these.

Finally, it exports 'AbstractAlgorithm' and 'AbstractResult' from 'MatrixBandwidth', and exports their various subtypes from the appropriate submodules. (At the same time, it removes the export of 'random_banded_matrix' from 'MatrixBandwidth'.)

(Fixes #138.)